+++ app/SearchApi/SearchAdapter.ICBC/Startup.cs
@@ -15,6 +15,7 @@ using SearchAdapter.ICBC.Infrastructure;
 using SearchAdapter.ICBC.SearchRequest;
 using SearchApi.Core.Configuration;
 using SearchApi.Core.Contracts;
+using SearchApi.Core.OpenTracing;
 
 namespace SearchAdapter.ICBC
 {
@@ -138,7 +139,9 @@ namespace SearchAdapter.ICBC
                     });
 
                     // Add Diagnostic context for tracing
-                    cfg.UseDiagnosticsActivity(new DiagnosticListener("open-tracing"));
+                    //cfg.UseDiagnosticsActivity(new DiagnosticListener("open-tracing"));
+
+                    cfg.PropagateOpenTracingContext();
 
                 }));
             });
