+++ app/SearchApi/SearchApi.Core/SearchApi.Core.csproj
@@ -6,6 +6,7 @@
 
   <ItemGroup>
     <PackageReference Include="MassTransit" Version="5.5.6" />
+    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="3.0.0" />
     <PackageReference Include="OpenTracing" Version="0.12.1" />
   </ItemGroup>
